🐛 [#148] 프로필 이미지 출력을 위한 수정

+++ backend/src/main/java/com/Devim/backend/domain/board/Board.java
@@ -14,7 +14,7 @@ public class Board {
 	private int boardTypeNo;
 	private String title;
 	private String boardContent;
-	private int writer;
+	private long writer;
 	private LocalDateTime createdDt;
 	private boolean deleteFlag;
 }
